Got rid of entity.js, folder inherits from file
