+++ plugins/Files/js/fileFactory.js
@@ -1,0 +1,28 @@
+'use strict';
+
+/*
+ * file factory module:
+ *   fileFactory, when called, makes a new file object
+ */
+
+// File object to make copy instances of
+const file = require('./file');
+
+// Factory to create instances of the file object
+function fileFactory(arg) {
+	// Files can be constructed from either a nickname or a status object
+	// returned from /renter/list||downloadqueue
+	var f = Object.create(file);
+	if (typeof arg === 'object'){
+		Object.assign(f, arg);
+		f.path = arg.nickname;
+	} else if (typeof arg === 'string') {
+		f.path = arg;
+	} else {
+		console.error('Unrecognized constructur argument: ', arguments);
+	}
+
+	return f;
+}
+
+module.exports = fileFactory;
